activities/requirements.txt
Flask
gunicorn
line-bot-sdk
markdown
beautifulsoup4
openai
google-genai
Pillow
Flask-Session